Complete migration of mock changes to new files
