+++ clients/admin-ui/src/features/data-discovery-and-detection/action-center/hooks/useDiscoveredInfrastructureSystemsTable.tsx
@@ -1,11 +1,20 @@
 import { AntEmpty as Empty } from "fidesui";
 import { useCallback, useMemo } from "react";
 
-import { UNCATEGORIZED_SEGMENT } from "~/features/common/nav/routes";
+import {
+  ACTION_CENTER_ROUTE,
+  UNCATEGORIZED_SEGMENT,
+} from "~/features/common/nav/routes";
 import { useAntTable, useTableState } from "~/features/common/table/hooks";
-import { SystemStagedResourcesAggregateRecord } from "~/types/api";
+import {
+  AlertLevel,
+  StagedResourceAPIResponse,
+  SystemStagedResourcesAggregateRecord,
+} from "~/types/api";
 
+import { useGetIdentityProviderMonitorResultsQuery } from "../../discovery-detection.slice";
 import { useGetDiscoveredSystemAggregateQuery } from "../action-center.slice";
+import { MONITOR_TYPES } from "../utils/getMonitorType";
 import useActionCenterTabs, {
   ActionCenterTabHash,
 } from "./useActionCenterTabs";
@@ -18,8 +27,8 @@ interface UseDiscoveredInfrastructureSystemsTableConfig {
 export const useDiscoveredInfrastructureSystemsTable = ({
   monitorId,
 }: UseDiscoveredInfrastructureSystemsTableConfig) => {
-  const { filterTabs, activeTab, onTabChange, activeParams } =
-    useActionCenterTabs();
+  // Check if this is an Okta app monitor
+  const isOktaApp = monitorId.toLowerCase().includes("okta");
 
   const tableState = useTableState<"name">({
     sorting: {
@@ -30,13 +39,108 @@ export const useDiscoveredInfrastructureSystemsTable = ({
   const { pageIndex, pageSize, searchQuery, updateSearch, resetState } =
     tableState;
 
-  const { data, isLoading, isFetching } = useGetDiscoveredSystemAggregateQuery({
-    key: monitorId,
-    page: pageIndex,
-    size: pageSize,
-    search: searchQuery,
-    ...activeParams,
-  });
+  // Get regular tabs for non-Okta monitors
+  const regularTabs = useActionCenterTabs();
+
+  // Use Identity Provider Monitor endpoint for Okta, otherwise use regular endpoint
+  const regularDataQuery = useGetDiscoveredSystemAggregateQuery(
+    {
+      key: monitorId,
+      page: pageIndex,
+      size: pageSize,
+      search: searchQuery,
+      ...(isOktaApp ? {} : regularTabs.activeParams),
+    },
+    {
+      skip: isOktaApp,
+    },
+  );
+
+  const oktaDataQuery = useGetIdentityProviderMonitorResultsQuery(
+    {
+      monitor_config_key: monitorId,
+      page: pageIndex,
+      size: pageSize,
+      search: searchQuery,
+    },
+    {
+      skip: !isOktaApp,
+    },
+  );
+
+  const {
+    data: regularData,
+    isLoading: regularIsLoading,
+    isFetching: regularIsFetching,
+  } = regularDataQuery;
+  const {
+    data: oktaData,
+    isLoading: oktaIsLoading,
+    isFetching: oktaIsFetching,
+  } = oktaDataQuery;
+
+  // Transform Okta data (StagedResourceAPIResponse[]) to SystemStagedResourcesAggregateRecord[]
+  // This is a temporary transformation until the BE provides the correct format
+  const transformedOktaData = useMemo(() => {
+    if (!oktaData?.items) {
+      return undefined;
+    }
+
+    return {
+      ...oktaData,
+      items: oktaData.items.map(
+        (
+          item: StagedResourceAPIResponse,
+        ): SystemStagedResourcesAggregateRecord => {
+          let consentStatus = null;
+          if (item.consent_breakdown) {
+            let alertLevel: AlertLevel;
+            if (item.consent_aggregated === "with_consent") {
+              alertLevel = AlertLevel.SUCCESS;
+            } else if (item.consent_aggregated === "without_consent") {
+              alertLevel = AlertLevel.ERROR;
+            } else {
+              alertLevel = AlertLevel.WARNING;
+            }
+            consentStatus = {
+              status: alertLevel,
+              message: item.consent_breakdown.message ?? "",
+            };
+          }
+
+          return {
+            id: item.urn,
+            name: item.name ?? null,
+            system_key: item.system_key ?? null,
+            data_uses: item.data_uses ?? [],
+            vendor_id: item.vendor_id ?? null,
+            total_updates: 1, // Okta apps are individual items, not aggregates
+            locations: item.locations ?? [],
+            domains: item.domain ? [item.domain] : [],
+            consent_status: consentStatus,
+            metadata: item.metadata ?? null,
+          };
+        },
+      ),
+    };
+  }, [oktaData]);
+
+  const data = isOktaApp ? transformedOktaData : regularData;
+  const isLoading = isOktaApp ? oktaIsLoading : regularIsLoading;
+  const isFetching = isOktaApp ? oktaIsFetching : regularIsFetching;
+
+  // For Okta, use simplified tabs (filters should be handled by BE in the future)
+  // For regular infrastructure, use standard tabs
+  const filterTabs = isOktaApp
+    ? [
+      {
+        label: "All Apps",
+        hash: "all",
+      },
+    ]
+    : regularTabs.filterTabs;
+  const activeTab: string = isOktaApp ? "all" : (regularTabs.activeTab ?? "");
+  const activeParams = isOktaApp ? {} : regularTabs.activeParams;
 
   // Helper function to generate consistent row keys
   const getRecordKey = useCallback(
@@ -75,15 +179,32 @@ export const useDiscoveredInfrastructureSystemsTable = ({
   const { selectedRows, resetSelections } = antTable;
 
   const handleTabChange = useCallback(
-    async (tab: ActionCenterTabHash) => {
-      await onTabChange(tab);
+    async (tab: string | ActionCenterTabHash) => {
+      if (isOktaApp) {
+        // For Okta, tab changes will be handled by BE filters in the future
+        // For now, we just reset state
+      } else {
+        await regularTabs.onTabChange(tab as ActionCenterTabHash);
+      }
       resetState();
       resetSelections();
     },
-    [onTabChange, resetState, resetSelections],
+    [isOktaApp, regularTabs, resetState, resetSelections],
   );
 
-  const { columns } = useDiscoveredInfrastructureSystemsColumns();
+  const rowClickUrl = useCallback(
+    (record: SystemStagedResourcesAggregateRecord) => {
+      const recordId = record.id ?? UNCATEGORIZED_SEGMENT;
+      const activeTabHash = activeTab ? `#${activeTab}` : "";
+      return `${ACTION_CENTER_ROUTE}/${MONITOR_TYPES.INFRASTRUCTURE}/${monitorId}/${recordId}${activeTabHash}`;
+    },
+    [monitorId, activeTab],
+  );
+
+  const { columns } = useDiscoveredInfrastructureSystemsColumns({
+    isOktaApp,
+    rowClickUrl,
+  });
 
   return {
     // Table state and data
