+++ clients/admin-ui/src/features/data-discovery-and-detection/action-center/hooks/useDiscoveredSystemAggregateTable.tsx
@@ -18,12 +18,9 @@ import {
   AlertLevel,
   ConsentAlertInfo,
   DiffStatus,
-  StagedResourceAPIResponse,
-  StagedResourceTypeValue,
   SystemStagedResourcesAggregateRecord,
 } from "~/types/api";
 
-import { useGetIdentityProviderMonitorResultsQuery } from "../../discovery-detection.slice";
 import {
   useAddMonitorResultSystemsMutation,
   useGetDiscoveredSystemAggregateQuery,
@@ -51,9 +48,6 @@ export const useDiscoveredSystemAggregateTable = ({
     ConsentAlertInfo | undefined
   >();
 
-  // Check if this is an Okta app monitor
-  const isOktaApp = monitorId.toLowerCase().includes("okta");
-
   const tableState = useTableState<DiscoveredSystemAggregateColumnKeys>({
     sorting: {
       validColumns: Object.values(DiscoveredSystemAggregateColumnKeys),
@@ -63,109 +57,16 @@ export const useDiscoveredSystemAggregateTable = ({
   const { pageIndex, pageSize, searchQuery, updateSearch, resetState } =
     tableState;
 
-  // Use appropriate tab management based on monitor type
-  const regularTabs = useActionCenterTabs();
+  const { filterTabs, activeTab, onTabChange, activeParams, actionsDisabled } =
+    useActionCenterTabs();
 
-  // Use Identity Provider Monitor endpoint for Okta, otherwise use regular endpoint
-  const regularDataQuery = useGetDiscoveredSystemAggregateQuery(
-    {
-      key: monitorId,
-      page: pageIndex,
-      size: pageSize,
-      search: searchQuery,
-      ...regularTabs.activeParams,
-    },
-    {
-      skip: isOktaApp,
-    },
-  );
-
-  const oktaDataQuery = useGetIdentityProviderMonitorResultsQuery(
-    {
-      monitor_config_key: monitorId,
-      page: pageIndex,
-      size: pageSize,
-      search: searchQuery,
-    },
-    {
-      skip: !isOktaApp,
-    },
-  );
-
-  const {
-    data: regularData,
-    isLoading: regularIsLoading,
-    isFetching: regularIsFetching,
-  } = regularDataQuery;
-  const {
-    data: oktaData,
-    isLoading: oktaIsLoading,
-    isFetching: oktaIsFetching,
-  } = oktaDataQuery;
-
-  // Transform Okta data (StagedResourceAPIResponse[]) to SystemStagedResourcesAggregateRecord[]
-  // This is a temporary transformation until the BE provides the correct format
-  const transformedOktaData = useMemo(() => {
-    if (!oktaData?.items) {
-      return undefined;
-    }
-
-    return {
-      ...oktaData,
-      items: oktaData.items.map(
-        (
-          item: StagedResourceAPIResponse,
-        ): SystemStagedResourcesAggregateRecord => {
-          let consentStatus: ConsentAlertInfo | null = null;
-          if (item.consent_breakdown) {
-            let alertLevel: AlertLevel;
-            if (item.consent_aggregated === "with_consent") {
-              alertLevel = AlertLevel.SUCCESS;
-            } else if (item.consent_aggregated === "without_consent") {
-              alertLevel = AlertLevel.ERROR;
-            } else {
-              alertLevel = AlertLevel.WARNING;
-            }
-            consentStatus = {
-              status: alertLevel,
-              message: item.consent_breakdown.message ?? "",
-            };
-          }
-
-          return {
-            id: item.urn,
-            name: item.name ?? null,
-            system_key: item.system_key ?? null,
-            data_uses: item.data_uses ?? [],
-            vendor_id: item.vendor_id ?? null,
-            total_updates: 1, // Okta apps are individual items, not aggregates
-            locations: item.locations ?? [],
-            domains: item.domain ? [item.domain] : [],
-            consent_status: consentStatus,
-            metadata: item.metadata ?? null,
-          };
-        },
-      ),
-    };
-  }, [oktaData]);
-
-  const data = isOktaApp ? transformedOktaData : regularData;
-  const isLoading = isOktaApp ? oktaIsLoading : regularIsLoading;
-  const isFetching = isOktaApp ? oktaIsFetching : regularIsFetching;
-
-  // For Okta, use simplified tabs (filters should be handled by BE in the future)
-  // For now, we'll use a basic structure that can be extended when BE supports filtering
-  const filterTabs = isOktaApp
-    ? [
-      {
-        label: "All Apps",
-        hash: "all",
-      },
-    ]
-    : regularTabs.filterTabs;
-  const activeTab: string = isOktaApp ? "all" : (regularTabs.activeTab ?? "");
-  const activeParams = isOktaApp ? {} : regularTabs.activeParams;
-  const actionsDisabled = isOktaApp ? false : regularTabs.actionsDisabled;
+  const { data, isLoading, isFetching } = useGetDiscoveredSystemAggregateQuery({
+    key: monitorId,
+    page: pageIndex,
+    size: pageSize,
+    search: searchQuery,
+    ...activeParams,
+  });
 
   const [addMonitorResultSystemsMutation, { isLoading: isAddingResults }] =
     useAddMonitorResultSystemsMutation();
@@ -183,14 +84,11 @@ export const useDiscoveredSystemAggregateTable = ({
 
   const rowClickUrl = useCallback(
     (record: SystemStagedResourcesAggregateRecord) => {
-      const monitorType = isOktaApp
-        ? MONITOR_TYPES.INFRASTRUCTURE
-        : MONITOR_TYPES.WEBSITE;
       const recordId = record.id ?? UNCATEGORIZED_SEGMENT;
       const activeTabHash = activeTab ? `#${activeTab}` : "";
-      return `${ACTION_CENTER_ROUTE}/${monitorType}/${monitorId}/${recordId}${activeTabHash}`;
+      return `${ACTION_CENTER_ROUTE}/${MONITOR_TYPES.WEBSITE}/${monitorId}/${recordId}${activeTabHash}`;
     },
-    [monitorId, activeTab, isOktaApp],
+    [monitorId, activeTab],
   );
 
   const antTableConfig = useMemo(
@@ -226,33 +124,26 @@ export const useDiscoveredSystemAggregateTable = ({
   const { selectedRows, resetSelections } = antTable;
 
   const handleTabChange = useCallback(
-    async (tab: string | ActionCenterTabHash) => {
+    async (tab: ActionCenterTabHash) => {
       setFirstPageConsentStatus(undefined);
-      if (isOktaApp) {
-        // For Okta, tab changes will be handled by BE filters in the future
-        // For now, we just reset state
-      } else {
-        await regularTabs.onTabChange(tab as ActionCenterTabHash);
-      }
+      await onTabChange(tab);
       resetState();
       resetSelections();
     },
-    [isOktaApp, regularTabs, resetState, resetSelections],
+    [onTabChange, resetState, resetSelections],
   );
 
   const { columns } = useDiscoveredSystemAggregateColumns({
     monitorId,
     onTabChange: handleTabChange,
     readonly: actionsDisabled,
-    allowIgnore: isOktaApp
-      ? false
-      : !("diff_status" in activeParams) ||
+    allowIgnore:
+      !("diff_status" in activeParams) ||
       !(activeParams as { diff_status: DiffStatus[] }).diff_status.includes(
         DiffStatus.MUTED,
       ),
     consentStatus: firstPageConsentStatus,
     rowClickUrl,
-    resourceType: isOktaApp ? StagedResourceTypeValue.OKTA_APP : undefined,
   });
 
   // Business logic effects
@@ -321,7 +212,7 @@ export const useDiscoveredSystemAggregateTable = ({
           SuccessToastContent(
             `${totalUpdates} assets have been ignored and will not appear in future scans.`,
             async () => {
-              await regularTabs.onTabChange(ActionCenterTabHash.IGNORED);
+              await onTabChange(ActionCenterTabHash.IGNORED);
             },
           ),
         ),
@@ -333,7 +224,7 @@ export const useDiscoveredSystemAggregateTable = ({
     ignoreMonitorResultSystemsMutation,
     monitorId,
     toast,
-    regularTabs,
+    onTabChange,
     resetSelections,
   ]);
 
@@ -374,8 +265,5 @@ export const useDiscoveredSystemAggregateTable = ({
     anyBulkActionIsLoading,
     isAddingResults,
     isIgnoringResults,
-
-    // Okta-specific functionality
-    isOktaApp,
   };
 };
